Fixed a lot of warnings in the engine that were clogging up the output window Most were deprecated function warnings many are also bogus u32 to f32 loss of data. I'll have to figure out how to turn those off with turning off f32 to u32 warnings.

+++ Engine/AnimationController.cpp
@@ -71,7 +71,7 @@ void AnimationPlayer::PlayAnimation(u32 animID, u32 frameOffset, f32 playSpeed)
 	
 	const u32 numFrames = m_pCurrAnim->endFrame-m_pCurrAnim->startFrame;
 	
-	m_currFrame = m_pCurrAnim->startFrame+frameOffset%numFrames;
+	m_currFrame = static_cast<f32>(m_pCurrAnim->startFrame+frameOffset%numFrames);
 }
 
 void AnimationPlayer::SetSpeed(f32 speed)
